document-policy/required-policy/document-policy.html

<!DOCTYPE html>
<html>
  <head>
    <title>Test advertised required document policy</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>  </head>
  <body>
    <h1>Test advertised required document policy</h1>
<script>
// The top-level document does not have a required document policy. Any
// required policy must come from the policy attribute of an iframe tag.

callbacks = {};

window.addEventListener('message', ev => {
  var id = ev.data.id;
  if (id && callbacks[id]) {
    callbacks[id](ev.data.requiredPolicy || null);
  }
});

async_test(t => {
  var iframe = document.createElement('iframe');
  iframe.src = "/document-policy/echo-policy.py?id=1";
  callbacks["1"] = t.step_func_done(result => {
    assert_equals(result, null);
  });
  document.body.appendChild(iframe);
}, "test advertised required document policy");

async_test(t => {
  var iframe = document.createElement('iframe');
  iframe.src = "/document-policy/echo-policy.py?id=2";
  iframe.policy = "unoptimized-lossless-images;bpp=4";
  callbacks["2"] = t.step_func_done(result => {
    assert_equals(result, "unoptimized-lossless-images;bpp=4");
  });
  document.body.appendChild(iframe);
}, "test advertised required document policy with iframe attribute");
    </script>
  </body>
</html>
